+++ ceammc/ext/src/fx/fx_flanger.h
@@ -1,42 +1,17 @@
 /* ------------------------------------------------------------
 name: "fx_flanger"
-Code generated with Faust 2.5.31 (https://faust.grame.fr)
+Code generated with Faust 2.8.5 (https://faust.grame.fr)
 Compilation options: cpp, -scal -ftz 0
 ------------------------------------------------------------ */
 
-#ifndef  __flanger_H__
-#define  __flanger_H__
+#ifndef  __fx_flanger_H__
+#define  __fx_flanger_H__
 
-/************************************************************************
- ************************************************************************
-    FAUST Architecture File
-    Copyright (C) 2006-2011 Albert Graef <Dr.Graef@t-online.de>
-    ---------------------------------------------------------------------
-    This program is free software; you can redistribute it and/or modify
-    it under the terms of the GNU Lesser General Public License as
-    published by the Free Software Foundation; either version 2.1 of the
-    License, or (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Lesser General Public License for more details.
-
-    You should have received a copy of the GNU Lesser General Public
-    License along with the GNU C Library; if not, write to the Free
-    Software Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA
-    02111-1307 USA.
- ************************************************************************
- ************************************************************************/
-
-/* Pd architecture file, written by Albert Graef <Dr.Graef@t-online.de>.
-   This was derived from minimal.cpp included in the Faust distribution.
-   Please note that this is to be compiled as a shared library, which is
-   then loaded dynamically by Pd as an external. */
-
-#include <math.h>
-#include <stdio.h>
-#include <stdlib.h>
+// FAUST Architecture File for ceammc::SoundExternal class
+#include <cmath>
+#include <cstdio>
+#include <cstdlib>
+#include <memory>
 #include <string>
 
 /************************************************************************
@@ -66,6 +41,7 @@ Compilation options: cpp, -scal -ftz 0
 #define __dsp__
 
 #include <string>
+#include <vector>
 
 #ifndef FAUSTFLOAT
 #define FAUSTFLOAT float
@@ -229,6 +205,9 @@ class dsp_factory {
         virtual std::string getName() = 0;
         virtual std::string getSHAKey() = 0;
         virtual std::string getDSPCode() = 0;
+        virtual std::string getCompileOptions() = 0;
+        virtual std::vector<std::string> getLibraryList() = 0;
+        virtual std::vector<std::string> getIncludePathnames() = 0;
     
         virtual dsp* createDSPInstance() = 0;
     
@@ -395,6 +374,7 @@ struct Meta
 #include <map>
 #include <string.h>
 #include <stdlib.h>
+#include <cstdlib>
 
 
 using std::max;
@@ -417,7 +397,7 @@ inline int int2pow2(int x)		{ int r = 0; while ((1<<r) < x) r++; return r; }
 inline long lopt(char* argv[], const char* name, long def)
 {
 	int	i;
-	for (i = 0; argv[i]; i++) if (!strcmp(argv[i], name)) return atoi(argv[i+1]);
+    for (i = 0; argv[i]; i++) if (!strcmp(argv[i], name)) return std::atoi(argv[i+1]);
 	return def;
 }
 
@@ -438,16 +418,9 @@ inline const char* lopts(char* argv[], const char* name, const char* def)
 #endif
 
 
-#include "ceammc_atomlist.h"
-#include <m_pd.h>
+#include "ceammc_faust.h"
 
-/******************************************************************************
-*******************************************************************************
-
-                                   VECTOR INTRINSICS
-
-*******************************************************************************
-*******************************************************************************/
+using namespace ceammc::faust;
 
 #ifdef FAUST_MACRO
 // clang-format off
@@ -457,31 +430,12 @@ inline const char* lopts(char* argv[], const char* name, const char* def)
 #define sym(name) xsym(name)
 #define xsym(name) #name
 
-/***************************************************************************
-   Pd UI interface
-***************************************************************************/
-
 // clang-format off
 #ifndef FAUST_MACRO
-struct flanger : public dsp {
+struct fx_flanger : public dsp {
 };
 #endif
 // clang-format on
-
-#include "ceammc_faust.h"
-using namespace ceammc::faust;
-
-/******************************************************************************
-*******************************************************************************
-
-                FAUST DSP
-
-*******************************************************************************
-*******************************************************************************/
-
-//----------------------------------------------------------------------------
-//  FAUST generated signal processor
-//----------------------------------------------------------------------------
 
 #ifdef FAUST_MACRO
 // clang-format off
@@ -489,28 +443,27 @@ using namespace ceammc::faust;
 #define FAUSTFLOAT float
 #endif 
 
+#include <algorithm>
 #include <cmath>
 #include <math.h>
 
 
 #ifndef FAUSTCLASS 
-#define FAUSTCLASS flanger
+#define FAUSTCLASS fx_flanger
 #endif
 #ifdef __APPLE__ 
 #define exp10f __exp10f
 #define exp10 __exp10
 #endif
 
-class flanger : public dsp {
+class fx_flanger : public dsp {
 	
  private:
 	
 	FAUSTFLOAT fCheckbox0;
-	FAUSTFLOAT fHslider0;
 	int iVec0[2];
-	float fRec1[2];
-	int IOTA;
-	float fVec1[4096];
+	FAUSTFLOAT fHslider0;
+	float fRec0[2];
 	int fSamplingFreq;
 	float fConst0;
 	float fConst1;
@@ -521,16 +474,22 @@ class flanger : public dsp {
 	float fRec3[2];
 	float fRec4[2];
 	float fRec2[2];
-	float fRec0[2];
-	FAUSTFLOAT fCheckbox1;
 	FAUSTFLOAT fHslider4;
 	float fRec5[2];
+	int IOTA;
+	float fVec1[4096];
+	float fRec1[2];
+	FAUSTFLOAT fCheckbox1;
+	FAUSTFLOAT fHslider5;
+	float fRec6[2];
 	
  public:
 	
 	void metadata(Meta* m) { 
 		m->declare("basics.lib/name", "Faust Basic Element Library");
 		m->declare("basics.lib/version", "0.0");
+		m->declare("ceammc.lib/name", "Ceammc PureData misc utils");
+		m->declare("ceammc.lib/version", "0.1.1");
 		m->declare("ceammc_ui.lib/name", "CEAMMC faust default UI elements");
 		m->declare("ceammc_ui.lib/version", "0.1.1");
 		m->declare("delays.lib/name", "Faust Delay Library");
@@ -599,7 +558,7 @@ class flanger : public dsp {
 	
 	virtual void instanceConstants(int samplingFreq) {
 		fSamplingFreq = samplingFreq;
-		fConst0 = min(192000.0f, max(1.0f, float(fSamplingFreq)));
+		fConst0 = std::min(192000.0f, std::max(1.0f, float(fSamplingFreq)));
 		fConst1 = (0.00100000005f * fConst0);
 		fConst2 = (6.28318548f / fConst0);
 		
@@ -607,12 +566,13 @@ class flanger : public dsp {
 	
 	virtual void instanceResetUserInterface() {
 		fCheckbox0 = FAUSTFLOAT(0.0f);
-		fHslider0 = FAUSTFLOAT(0.0f);
+		fHslider0 = FAUSTFLOAT(1.0f);
 		fHslider1 = FAUSTFLOAT(1.0f);
 		fHslider2 = FAUSTFLOAT(10.0f);
 		fHslider3 = FAUSTFLOAT(0.5f);
+		fHslider4 = FAUSTFLOAT(0.0f);
 		fCheckbox1 = FAUSTFLOAT(0.0f);
-		fHslider4 = FAUSTFLOAT(1.0f);
+		fHslider5 = FAUSTFLOAT(1.0f);
 		
 	}
 	
@@ -622,32 +582,36 @@ class flanger : public dsp {
 			
 		}
 		for (int l1 = 0; (l1 < 2); l1 = (l1 + 1)) {
-			fRec1[l1] = 0.0f;
+			fRec0[l1] = 0.0f;
 			
 		}
-		IOTA = 0;
-		for (int l2 = 0; (l2 < 4096); l2 = (l2 + 1)) {
-			fVec1[l2] = 0.0f;
+		for (int l2 = 0; (l2 < 2); l2 = (l2 + 1)) {
+			fRec3[l2] = 0.0f;
 			
 		}
 		for (int l3 = 0; (l3 < 2); l3 = (l3 + 1)) {
-			fRec3[l3] = 0.0f;
+			fRec4[l3] = 0.0f;
 			
 		}
 		for (int l4 = 0; (l4 < 2); l4 = (l4 + 1)) {
-			fRec4[l4] = 0.0f;
+			fRec2[l4] = 0.0f;
 			
 		}
 		for (int l5 = 0; (l5 < 2); l5 = (l5 + 1)) {
-			fRec2[l5] = 0.0f;
+			fRec5[l5] = 0.0f;
 			
 		}
-		for (int l6 = 0; (l6 < 2); l6 = (l6 + 1)) {
-			fRec0[l6] = 0.0f;
+		IOTA = 0;
+		for (int l6 = 0; (l6 < 4096); l6 = (l6 + 1)) {
+			fVec1[l6] = 0.0f;
 			
 		}
 		for (int l7 = 0; (l7 < 2); l7 = (l7 + 1)) {
-			fRec5[l7] = 0.0f;
+			fRec1[l7] = 0.0f;
+			
+		}
+		for (int l8 = 0; (l8 < 2); l8 = (l8 + 1)) {
+			fRec6[l8] = 0.0f;
 			
 		}
 		
@@ -663,8 +627,8 @@ class flanger : public dsp {
 		instanceClear();
 	}
 	
-	virtual flanger* clone() {
-		return new flanger();
+	virtual fx_flanger* clone() {
+		return new fx_flanger();
 	}
 	virtual int getSampleRate() {
 		return fSamplingFreq;
@@ -677,10 +641,12 @@ class flanger : public dsp {
 		ui_interface->declare(&fHslider2, "style", "knob");
 		ui_interface->declare(&fHslider2, "unit", "ms");
 		ui_interface->addHorizontalSlider("delay", &fHslider2, 10.0f, 0.0f, 20.0f, 0.00100000005f);
-		ui_interface->declare(&fHslider4, "style", "knob");
-		ui_interface->addHorizontalSlider("drywet", &fHslider4, 1.0f, 0.0f, 1.0f, 0.00100000005f);
 		ui_interface->declare(&fHslider0, "style", "knob");
-		ui_interface->addHorizontalSlider("feedback", &fHslider0, 0.0f, -0.999000013f, 0.999000013f, 0.00100000005f);
+		ui_interface->addHorizontalSlider("drywet", &fHslider0, 1.0f, 0.0f, 1.0f, 0.00999999978f);
+		ui_interface->declare(&fHslider5, "style", "knob");
+		ui_interface->addHorizontalSlider("drywet", &fHslider5, 1.0f, 0.0f, 1.0f, 0.00100000005f);
+		ui_interface->declare(&fHslider4, "style", "knob");
+		ui_interface->addHorizontalSlider("feedback", &fHslider4, 0.0f, -0.999000013f, 0.999000013f, 0.00100000005f);
 		ui_interface->addCheckButton("invert", &fCheckbox1);
 		ui_interface->declare(&fHslider1, "style", "knob");
 		ui_interface->declare(&fHslider1, "unit", "ms");
@@ -700,34 +666,37 @@ class flanger : public dsp {
 		float fSlow2 = (0.00100000005f * float(fHslider1));
 		float fSlow3 = (0.000500000024f * float(fHslider2));
 		float fSlow4 = (fConst2 * float(fHslider3));
-		float fSlow5 = sinf(fSlow4);
-		float fSlow6 = cosf(fSlow4);
+		float fSlow5 = std::sin(fSlow4);
+		float fSlow6 = std::cos(fSlow4);
 		float fSlow7 = (0.0f - fSlow5);
-		int iSlow8 = int(float(fCheckbox1));
-		float fSlow9 = (0.00100000005f * float(fHslider4));
+		float fSlow8 = (0.00100000005f * float(fHslider4));
+		int iSlow9 = int(float(fCheckbox1));
+		float fSlow10 = (0.00100000005f * float(fHslider5));
 		for (int i = 0; (i < count); i = (i + 1)) {
 			float fTemp0 = float(input0[i]);
 			float fTemp1 = (iSlow0?0.0f:fTemp0);
 			iVec0[0] = 1;
-			fRec1[0] = (fSlow1 + (0.999000013f * fRec1[1]));
-			float fTemp2 = ((fRec1[0] * fRec0[1]) - fTemp1);
-			fVec1[(IOTA & 4095)] = fTemp2;
+			fRec0[0] = (fSlow1 + (0.999000013f * fRec0[1]));
 			fRec3[0] = ((fSlow5 * fRec4[1]) + (fSlow6 * fRec3[1]));
 			fRec4[0] = (((fSlow6 * fRec4[1]) + (fSlow7 * fRec3[1])) + float((1 - iVec0[1])));
 			fRec2[0] = ((0.999000013f * fRec2[1]) + (fConst1 * (fSlow2 + (fSlow3 * (fRec3[0] + 1.0f)))));
-			int iTemp3 = int(fRec2[0]);
-			float fTemp4 = floorf(fRec2[0]);
-			fRec0[0] = ((fVec1[((IOTA - min(2049, max(0, (iTemp3 + 1)))) & 4095)] * (fRec2[0] - fTemp4)) + ((fTemp4 + (1.0f - fRec2[0])) * fVec1[((IOTA - min(2049, max(0, iTemp3))) & 4095)]));
-			fRec5[0] = (fSlow9 + (0.999000013f * fRec5[1]));
-			output0[i] = FAUSTFLOAT((iSlow0?fTemp0:(0.5f * (fTemp1 + (fRec0[0] * (iSlow8?(0.0f - fRec5[0]):fRec5[0]))))));
+			float fTemp2 = std::floor(fRec2[0]);
+			fRec5[0] = (fSlow8 + (0.999000013f * fRec5[1]));
+			float fTemp3 = ((fRec1[1] * fRec5[0]) - fTemp1);
+			fVec1[(IOTA & 4095)] = fTemp3;
+			int iTemp4 = int(fRec2[0]);
+			fRec1[0] = (((fTemp2 + (1.0f - fRec2[0])) * fVec1[((IOTA - std::min(2049, std::max(0, iTemp4))) & 4095)]) + ((fRec2[0] - fTemp2) * fVec1[((IOTA - std::min(2049, std::max(0, (iTemp4 + 1)))) & 4095)]));
+			fRec6[0] = (fSlow10 + (0.999000013f * fRec6[1]));
+			output0[i] = FAUSTFLOAT((iSlow0?fTemp0:((fTemp1 * (1.0f - fRec0[0])) + (0.5f * (fRec0[0] * (fTemp1 + (fRec1[0] * (iSlow9?(0.0f - fRec6[0]):fRec6[0]))))))));
 			iVec0[1] = iVec0[0];
-			fRec1[1] = fRec1[0];
-			IOTA = (IOTA + 1);
+			fRec0[1] = fRec0[0];
 			fRec3[1] = fRec3[0];
 			fRec4[1] = fRec4[0];
 			fRec2[1] = fRec2[0];
-			fRec0[1] = fRec0[0];
 			fRec5[1] = fRec5[0];
+			IOTA = (IOTA + 1);
+			fRec1[1] = fRec1[0];
+			fRec6[1] = fRec6[0];
 			
 		}
 		
@@ -738,586 +707,22 @@ class flanger : public dsp {
 // clang-format on
 #endif
 
-// clang-format off
-#define faust_setup(name) xfaust_setup(name)
-#define xfaust_setup(name) name##_tilde_setup(void)
-// time for "active" toggle xfades in secs
-#define XFADE_TIME 0.1f
-static t_class* flanger_faust_class;
-#define FAUST_EXT t_faust_flanger
-#define FAUST_EXT_CLASS flanger_faust_class
-// clang-format on
-
-struct t_faust_flanger {
-    t_object x_obj;
-#ifdef __MINGW32__
-    /* This seems to be necessary as some as yet undetermined Pd routine seems
-     to write past the end of x_obj on Windows. */
-    int fence; /* dummy field (not used) */
-#endif
-    flanger* dsp;
-    PdUI<UI>* ui;
-    int active, xfade, n_xfade, rate, n_in, n_out;
-    t_sample **inputs, **outputs, **buf;
-    t_outlet* out;
-    t_sample f;
+    template <class T>
+    struct _fx_flanger_UI : public UI {
+    static std::string name;
 };
 
-static inline void zero_samples(int k, int n, t_sample** out)
-{
-    for (int i = 0; i < k; i++)
-#ifdef __STDC_IEC_559__
-        /* IEC 559 a.k.a. IEEE 754 floats can be initialized faster like this */
-        memset(out[i], 0, n * sizeof(t_sample));
-#else
-        for (int j = 0; j < n; j++)
-            out[i][j] = 0.0f;
-#endif
-}
+template <class T>
+std::string _fx_flanger_UI<T>::name(sym(fx_flanger));
 
-static inline void copy_samples(int k, int n, t_sample** out, t_sample** in)
-{
-    for (int i = 0; i < k; i++)
-        memcpy(out[i], in[i], n * sizeof(t_sample));
-}
+typedef _fx_flanger_UI<fx_flanger> fx_flanger_UI;
 
-static t_int* faust_perform(t_int* w)
-{
-    t_faust_flanger* x = reinterpret_cast<t_faust_flanger*>(w[1]);
-    int n = static_cast<int>(w[2]);
-    if (!x->dsp || !x->buf)
-        return (w + 3);
-
-    AVOIDDENORMALS;
-    if (x->xfade > 0) {
-        float d = 1.0f / x->n_xfade, f = (x->xfade--) * d;
-        d = d / n;
-        x->dsp->compute(n, x->inputs, x->buf);
-        if (x->active) {
-            if (x->n_in == x->n_out) {
-                /* xfade inputs -> buf */
-                for (int j = 0; j < n; j++, f -= d) {
-                    for (int i = 0; i < x->n_out; i++)
-                        x->outputs[i][j] = f * x->inputs[i][j] + (1.0f - f) * x->buf[i][j];
-                }
-            } else {
-                /* xfade 0 -> buf */
-                for (int j = 0; j < n; j++, f -= d) {
-                    for (int i = 0; i < x->n_out; i++)
-                        x->outputs[i][j] = (1.0f - f) * x->buf[i][j];
-                }
-            }
-        } else if (x->n_in == x->n_out) {
-            /* xfade buf -> inputs */
-            for (int j = 0; j < n; j++, f -= d) {
-                for (int i = 0; i < x->n_out; i++)
-                    x->outputs[i][j] = f * x->buf[i][j] + (1.0f - f) * x->inputs[i][j];
-            }
-        } else {
-            /* xfade buf -> 0 */
-            for (int j = 0; j < n; j++, f -= d) {
-                for (int i = 0; i < x->n_out; i++)
-                    x->outputs[i][j] = f * x->buf[i][j];
-            }
-        }
-    } else if (x->active) {
-        x->dsp->compute(n, x->inputs, x->buf);
-        copy_samples(x->n_out, n, x->outputs, x->buf);
-    } else if (x->n_in == x->n_out) {
-        copy_samples(x->n_out, n, x->buf, x->inputs);
-        copy_samples(x->n_out, n, x->outputs, x->buf);
-    } else
-        zero_samples(x->n_out, n, x->outputs);
-
-    return (w + 3);
-}
-
-static void flanger_faust_dsp(t_faust_flanger* x, t_signal** sp)
-{
-    const int n = sp[0]->s_n;
-    const int sr = static_cast<int>(sp[0]->s_sr);
-
-    if (x->rate <= 0) {
-        /* default sample rate is whatever Pd tells us */
-        PdUI<UI>* ui = x->ui;
-        std::vector<FAUSTFLOAT> z = ui->uiValues();
-        /* set the proper sample rate; this requires reinitializing the dsp */
-        x->rate = sr;
-        x->dsp->init(sr);
-        ui->setUIValues(z);
-    }
-    if (n > 0)
-        x->n_xfade = static_cast<int>(x->rate * XFADE_TIME / n);
-
-    dsp_add(faust_perform, 2, x, n);
-
-    for (int i = 0; i < x->n_in; i++)
-        x->inputs[i] = sp[i]->s_vec;
-
-    for (int i = 0; i < x->n_out; i++)
-        x->outputs[i] = sp[x->n_in + i]->s_vec;
-
-    if (x->buf != NULL) {
-        for (int i = 0; i < x->n_out; i++) {
-            x->buf[i] = static_cast<t_sample*>(malloc(n * sizeof(t_sample)));
-            if (x->buf[i] == NULL) {
-                for (int j = 0; j < i; j++)
-                    free(x->buf[j]);
-                free(x->buf);
-                x->buf = NULL;
-                break;
-            }
-        }
-    }
-}
-
-static void flanger_dump_to_console(t_faust_flanger* x)
-{
-    t_object* xobj = &x->x_obj;
-    t_class* xc = xobj->te_pd;
-    const char* name = class_getname(xc);
-
-    // print xlets
-    post("[%s] inlets: %i", name, x->dsp->getNumInputs());
-    int info_outlet = (x->out == 0) ? 0 : 1;
-    post("[%s] outlets: %i", name, x->dsp->getNumOutputs() + info_outlet);
-
-    // print properties
-    for (size_t i = 0; i < x->ui->uiCount(); i++) {
-        UIElement* el = x->ui->uiAt(i);
-        post("[%s] property: %s = %g", name, el->setPropertySym()->s_name, static_cast<double>(el->value()));
-    }
-}
-
-static void flanger_faust_any(t_faust_flanger* x, t_symbol* s, int argc, t_atom* argv)
-{
-    if (!x->dsp)
-        return;
-
-    PdUI<UI>* ui = x->ui;
-    if (s == &s_bang) {
-        ui->dumpUI(x->out);
-    } else if (isGetAllProperties(s)) {
-        ui->outputAllProperties(x->out);
-    } else if (isGetProperty(s)) {
-        ui->outputProperty(s, x->out);
-    } else if (isSetProperty(s)) {
-        ui->setProperty(s, argc, argv);
-    } else {
-        const char* label = s->s_name;
-        int count = 0;
-        for (size_t i = 0; i < ui->uiCount(); i++) {
-            if (ui->uiAt(i)->pathcmp(label)) {
-                if (argc == 0) {
-                    ui->uiAt(i)->outputValue(x->out);
-                    ++count;
-                } else if (argc == 1 && (argv[0].a_type == A_FLOAT || argv[0].a_type == A_DEFFLOAT)) {
-                    float f = atom_getfloat(argv);
-                    UIElement* el = ui->uiAt(i);
-                    el->setValue(f);
-                    ++count;
-                } else
-                    pd_error(x, "[ceammc] %s: bad control argument: %s", ui->fullName().c_str(), label);
-            }
-        }
-
-        if (count == 0 && strcmp(label, "active") == 0) {
-            if (argc == 0) {
-                t_atom arg;
-                SETFLOAT(&arg, (float)x->active);
-                if (x->out) {
-                    outlet_anything(x->out, gensym("active"), 1, &arg);
-                }
-            } else if (argc == 1 && (argv[0].a_type == A_FLOAT || argv[0].a_type == A_DEFFLOAT)) {
-                float f = atom_getfloat(argv);
-                x->active = (int)f;
-                x->xfade = x->n_xfade;
-            }
-        }
-    }
-}
-
-static void faust_free_dsp(t_faust_flanger* x)
-{
-    delete x->dsp;
-    x->dsp = NULL;
-}
-
-static void faust_free_ui(t_faust_flanger* x)
-{
-    delete x->ui;
-    x->ui = NULL;
-}
-
-static void faust_free_inputs(t_faust_flanger* x)
-{
-    if (x->inputs)
-        free(x->inputs);
-    x->inputs = NULL;
-}
-
-static void faust_free_outputs(t_faust_flanger* x)
-{
-    if (x->outputs)
-        free(x->outputs);
-    x->outputs = NULL;
-}
-
-static void faust_free_buf(t_faust_flanger* x)
-{
-    if (x->buf) {
-        for (int i = 0; i < x->n_out; i++) {
-            if (x->buf[i])
-                free(x->buf[i]);
-        }
-
-        free(x->buf);
-    }
-}
-
-static void flanger_faust_free(t_faust_flanger* x)
-{
-    faust_free_dsp(x);
-    faust_free_ui(x);
-    faust_free_inputs(x);
-    faust_free_outputs(x);
-    faust_free_buf(x);
-}
-
-static bool faust_init_inputs(t_faust_flanger* x)
-{
-    x->inputs = NULL;
-    x->n_in = x->dsp->getNumInputs();
-
-    if (x->n_in > 0) {
-        x->inputs = static_cast<t_sample**>(calloc(x->n_in, sizeof(t_sample*)));
-
-        if (x->inputs == NULL) {
-            error("[%s] faust_init_inputs failed", sym(flanger));
-            return false;
-        }
-    }
-
-    // creating sound inlets (except the first one!)
-    for (int i = 0; i < (x->n_in - 1); i++) {
-        inlet_new(&x->x_obj, &x->x_obj.ob_pd, &s_signal, &s_signal);
-    }
-
-    return true;
-}
-
-static bool faust_init_outputs(t_faust_flanger* x, bool info_outlet)
-{
-    x->outputs = NULL;
-    x->buf = NULL;
-
-    x->n_out = x->dsp->getNumOutputs();
-
-    if (x->n_out > 0) {
-        x->outputs = static_cast<t_sample**>(calloc(x->n_out, sizeof(t_sample*)));
-        if (x->outputs == NULL) {
-            error("[%s] faust_init_outputs failed", sym(flanger));
-            return false;
-        }
-
-        x->buf = static_cast<t_sample**>(calloc(x->n_out, sizeof(t_sample*)));
-        if (x->buf == NULL) {
-            error("[%s] faust_init_outputs failed", sym(flanger));
-            faust_free_outputs(x);
-            return false;
-        }
-
-        for (int i = 0; i < x->n_out; i++)
-            x->buf[i] = NULL;
-    }
-
-    // creating sound outlets
-    for (int i = 0; i < x->n_out; i++) {
-        outlet_new(&x->x_obj, &s_signal);
-    }
-
-    // control outlet
-    if (info_outlet)
-        x->out = outlet_new(&x->x_obj, 0);
-    else
-        x->out = 0;
-
-    return true;
-}
-
-static bool faust_new_internal(t_faust_flanger* x, const std::string& objId = "", bool info_outlet = true)
-{
-    int sr = 44100;
-    x->active = 1;
-    x->xfade = 0;
-    x->rate = sr;
-    x->n_xfade = static_cast<int>(sr * XFADE_TIME / 64);
-
-    x->dsp = new flanger();
-    x->ui = new PdUI<UI>(sym(flanger), objId);
-
-    if (!faust_init_inputs(x)) {
-        flanger_faust_free(x);
-        return false;
-    }
-
-    if (!faust_init_outputs(x, info_outlet)) {
-        flanger_faust_free(x);
-        return false;
-    }
-
-    x->dsp->init(sr);
-    x->dsp->buildUserInterface(x->ui);
-
-    return true;
-}
-
-/**
- * find nth element that satisfies given predicate
- * @param first - first element of sequence
- * @param last - pointer behind last element of sequence
- * @param Nth - searched element index
- * @param pred - predicate
- * @return pointer to found element or pointer to last, if not found
- */
-template <class InputIterator, class NthOccurence, class UnaryPredicate>
-InputIterator find_nth_if(InputIterator first, InputIterator last, NthOccurence Nth, UnaryPredicate pred)
-{
-    if (Nth > 0) {
-        while (first != last) {
-            if (pred(*first))
-                if (!--Nth)
-                    return first;
-            ++first;
-        }
-    }
-    return last;
-}
-
-/**
- * @return true if given atom is a float
- */
-static bool atom_is_float(const t_atom& a)
-{
-    switch (a.a_type) {
-    case A_FLOAT:
-    case A_DEFFLOAT:
-        return true;
-    default:
-        return false;
-    }
-}
-
-/**
- * @return true if given atom is a symbol
- */
-static bool atom_is_symbol(const t_atom& a)
-{
-    switch (a.a_type) {
-    case A_DEFSYMBOL:
-    case A_SYMBOL:
-        return true;
-    default:
-        return false;
-    }
-}
-
-/**
- * @return true if given atom is a property
- */
-static bool atom_is_property(const t_atom& a)
-{
-    switch (a.a_type) {
-    case A_DEFSYMBOL:
-    case A_SYMBOL:
-        return a.a_w.w_symbol->s_name[0] == '@';
-    default:
-        return false;
-    }
-}
-
-/**
- * @brief find nth float in argument list. (arguments can be mixed)
- * @param argc argument count
- * @param argv pointer to argument vector
- * @param nth find position. nth should be > 0!
- * @param dest destination to write value
- * @return true if argument at given position was found, otherwise false
- */
-static bool get_nth_float_arg(int argc, t_atom* argv, int nth, t_float* dest)
-{
-    t_atom* last = argv + argc;
-    t_atom* res = find_nth_if(argv, last, nth, atom_is_float);
-    if (last == res)
-        return false;
-
-    *dest = atom_getfloat(res);
-    return true;
-}
-
-/**
- * @brief find nth symbol in argument list. (arguments can be mixed)
- * @param argc argument count
- * @param argv pointer to argument vector
- * @param nth find position. nth should be > 0!
- * @param dest destination to write found argument value
- * @return true if argument at given position was found, otherwise false
- */
-static bool get_nth_symbol_arg(int argc, t_atom* argv, int nth, const char** dest)
-{
-    t_atom* last = argv + argc;
-    t_atom* res = find_nth_if(argv, last, nth, atom_is_symbol);
-    if (last == res)
-        return false;
-
-    t_symbol* s = atom_getsymbol(res);
-    *dest = s->s_name;
-    return true;
-}
-
-class PdArgParser {
-    t_faust_flanger* x_;
-    int argc_;
-    t_atom* argv_;
-    bool control_outlet_;
-
+class faust_fx_flanger_tilde : public FaustExternal<fx_flanger, fx_flanger_UI> {
 public:
-    /**
-     * @brief FaustArgParser
-     * @param x pointer to faust class
-     * @param argc arguments count
-     * @param argv pointer to argument vector
-     */
-    PdArgParser(t_faust_flanger* x, int argc, t_atom* argv, bool info_outlet = true)
-        : x_(x)
-        , argc_(0)
-        , argv_(argv)
-        , control_outlet_(info_outlet)
+    faust_fx_flanger_tilde(const ceammc::PdArgs& args)
+        : FaustExternal(args)
     {
-        const char* id = NULL;
-        std::string objId;
-
-        int first_prop_idx = argc;
-        for(int i = 0; i < argc; i++) {
-            if(atom_is_property(argv[i]))
-                first_prop_idx = i;
-        }
-
-        // store argument count (without properties)
-        argc_ = first_prop_idx;
-
-        if (get_nth_symbol_arg(argc_, argv_, 1, &id))
-            objId = id;
-
-        // init error
-        if (!faust_new_internal(x, objId, control_outlet_)) {
-            this->x_ = NULL;
-        }
-
-        // process properties
-        std::deque<ceammc::AtomList> props = ceammc::AtomList(argc, argv).properties();
-        for (size_t i = 0; i < props.size(); i++) {
-            ceammc::AtomList& p = props[i];
-            // skip empty property
-            if (p.size() < 2)
-                continue;
-
-            t_atom* data = p.toPdData() + 1;
-            this->x_->ui->setProperty(p[0].asSymbol(), p.size() - 1, data);
-        }
-    }
-
-    /**
-     * @brief initFloatArg
-     * @param name argument name
-     * @param pos argument position among of @bold float(!) arguments. Position starts from @bold 1(!).
-     * to select first argument - pass 1.
-     */
-    void initFloatArg(const char* name, int pos)
-    {
-        // object was not created
-        if (!this->x_)
-            return;
-
-        t_float v = 0.0;
-        if (get_nth_float_arg(this->argc_, this->argv_, pos, &v)) {
-            UIElement* el = this->x_->ui->findElementByLabel(name);
-            if (!el) {
-                post("invalid UI element name: %s", name);
-                return;
-            }
-
-            el->setValue(v, true);
-        }
-    }
-
-    /**
-     * @brief send creation argument to first signal inlet
-     * @param name argument name
-     * @param pos argument position among of @bold float(!) arguments. Position starts from @bold 1(!).
-     * to select first argument - pass 1.
-     */
-    void signalFloatArg(const char* /*name*/, int pos)
-    {
-        // object was not created
-        if (!this->x_)
-            return;
-
-        t_float arg = 0;
-        if (get_nth_float_arg(this->argc_, this->argv_, pos, &arg))
-            pd_float(reinterpret_cast<t_pd*>(this->x_), arg);
-    }
-
-    t_faust_flanger* pd_obj()
-    {
-        return this->x_;
     }
 };
-
-static void* flanger_faust_new(t_symbol* s, int argc, t_atom* argv);
-
-static void internal_setup(t_symbol* s, bool soundIn = true)
-{
-    flanger_faust_class = class_new(s, reinterpret_cast<t_newmethod>(flanger_faust_new),
-        reinterpret_cast<t_method>(flanger_faust_free),
-        sizeof(t_faust_flanger),
-        CLASS_DEFAULT,
-        A_GIMME, A_NULL);
-
-    if (soundIn) {
-        class_addmethod(flanger_faust_class, nullfn, &s_signal, A_NULL);
-        CLASS_MAINSIGNALIN(flanger_faust_class, t_faust_flanger, f);
-    }
-
-    class_addmethod(flanger_faust_class, reinterpret_cast<t_method>(flanger_faust_dsp), gensym("dsp"), A_NULL);
-    class_addmethod(flanger_faust_class, reinterpret_cast<t_method>(flanger_dump_to_console), gensym("dump"), A_NULL);
-    class_addanything(flanger_faust_class, flanger_faust_any);
-}
-
-#define EXTERNAL_NEW void* flanger_faust_new(t_symbol*, int argc, t_atom* argv)
-
-#define EXTERNAL_SIMPLE_NEW()                                                           \
-    static void* flanger_faust_new(t_symbol*, int argc, t_atom* argv)                     \
-    {                                                                                   \
-        t_faust_flanger* x = reinterpret_cast<t_faust_flanger*>(pd_new(flanger_faust_class)); \
-        PdArgParser p(x, argc, argv, false);                                            \
-        return p.pd_obj();                                                              \
-    }
-
-#define EXTERNAL_SETUP(MOD)                        \
-    extern "C" void setup_##MOD##0x2eflanger_tilde() \
-    {                                              \
-        internal_setup(gensym(#MOD ".flanger~"));    \
-    }
-
-#define EXTERNAL_SETUP_NO_IN(MOD)                      \
-    extern "C" void setup_##MOD##0x2eflanger_tilde()     \
-    {                                                  \
-        internal_setup(gensym(#MOD ".flanger~"), false); \
-    }
-
-#define SIMPLE_EXTERNAL(MOD) \
-    EXTERNAL_SIMPLE_NEW();   \
-    EXTERNAL_SETUP(MOD);
 
 #endif