+++ ceammc/ext/src/fx/mod_fx.cpp
@@ -4,7 +4,6 @@
 extern "C" void setup_fx0x2edistortion_tilde();
 extern "C" void setup_fx0x2edrone_box_tilde();
 extern "C" void setup_fx0x2eecho_tilde();
-extern "C" void setup_fx0x2eflanger_tilde();
 extern "C" void setup_fx0x2efreeverb2_tilde();
 extern "C" void setup_fx0x2efreeverb_tilde();
 extern "C" void setup_fx0x2efreqshift_tilde();
@@ -15,12 +14,13 @@ extern "C" void setup_fx0x2ewahwah_tilde();
 extern "C" void setup_fx0x2ezita_rev1_tilde();
 extern "C" void setup_fx0x2esdelay_tilde();
 
+extern void setup_fx_flanger_tilde();
+
 void ceammc_fx_setup()
 {
     setup_fx0x2edistortion_tilde();
     setup_fx0x2edrone_box_tilde();
     setup_fx0x2eecho_tilde();
-    setup_fx0x2eflanger_tilde();
     setup_fx0x2efreeverb2_tilde();
     setup_fx0x2efreeverb_tilde();
     setup_fx0x2efreqshift_tilde();
@@ -30,5 +30,7 @@ void ceammc_fx_setup()
     setup_fx0x2ewahwah_tilde();
     setup_fx0x2ezita_rev1_tilde();
     setup_fx0x2esdelay_tilde();
+
+    setup_fx_flanger_tilde();
     setup_fx_looper();
 }